feat: combine program & command
