+++ lib/command.js
@@ -1,11 +1,14 @@
 'use strict';
 
 const debug = require('debug')('common-bin');
+const co = require('co');
 const yargs = require('yargs');
 const helper = require('./helper');
 const assert = require('assert');
 const fs = require('fs');
 const path = require('path');
+const DISPATCH = Symbol('Command#dispatch');
+const pkg = require('../package.json');
 
 class Command {
   constructor() {
@@ -20,6 +23,9 @@ class Command {
      * @type {Object}
      */
     this.yargs = yargs;
+
+    this.version = pkg.version;
+    this.usage = '';
 
     /**
      * command name, just `start` or advance usage `get <source> [proxy]`
@@ -118,6 +124,54 @@ class Command {
   }
 
   /**
+   * start point of bin process
+   */
+  start() {
+    this.name = this.name || pkg.name;
+    this.yargs
+      .wrap(120)
+      .version(this.version)
+      .help()
+      .alias('h', 'help')
+      .group([ 'help', 'version' ], 'Global Options:')
+      // .strict()
+      // .demandCommand(1)
+      // .recommendCommands()
+      // .epilogue(`for more information, find our manual at ${pkg.homepage}`)
+      .usage(this.usage)
+      .options(this.options);
+
+    // register to yargs
+    this.commandMapping.forEach(command => {
+      debug('register command `%s`', command.name);
+      this.yargs.command(command.toYargs());
+    });
+
+    const context = {
+      // { _: [ 'init' ], help: false, baseDir: 1234, '$0': 'bin.js' }
+      argv: this.yargs.argv,
+      cwd: process.cwd(),
+      rawArgv: process.argv.slice(3),
+    };
+
+
+    co(function* () {
+      yield this[DISPATCH](context);
+    }.bind(this)).catch(this.errorHandler.bind(this));
+  }
+
+  /**
+   * default error hander
+   * @param {Error} err - error object
+   * @protected
+   */
+  errorHandler(err) {
+    console.error('⚠️ [%s] run command %s with %j got error: %s', this.name, process.argv[2] || '', process.argv.slice(3), err.message);
+    console.log(err.stack);
+    process.exit(1);
+  }
+
+  /**
    * dispatch command, either `subCommand.exec` or `this.run`
    * @param {Object} context - context object
    * @param {String} context.cwd - process.cwd()
@@ -125,7 +179,7 @@ class Command {
    * @param {Array} context.rawArgv - the raw argv, `[ "--baseDir=simple" ]`
    * @private
    */
-  * exec(context) {
+  * [DISPATCH](context) {
     let commandName = context.argv._[0];
     if (this.aliasMapping.has(commandName)) {
       commandName = this.aliasMapping.get(commandName);
@@ -136,7 +190,7 @@ class Command {
       // exec sub command
       debug('dispatch to subcommand `%s` -> `%s`', this.name, command.name);
       context.argv._ = context.argv._.slice(1);
-      yield this.helper.callFn(command.exec.bind(command), [ context ]);
+      yield this.helper.callFn(command[DISPATCH].bind(command), [ context ]);
     } else if (this.run) {
       debug('exec command `%s`', this.name);
       yield this.helper.callFn(this.run.bind(this), [ context ]);
