+++ test/fixtures/async-bin/index.js
