+++ test/fixtures/my-bin/lib/command/sub/not-register.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../../../..').Command;
+const Command = require('../../../../../..');
 
 class NoCommand extends Command {
   constructor() {
