+++ test/fixtures/my-bin/lib/cov_command.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../..').Command;
+const Command = require('../../../..');
 
 class CovCommand extends Command {
   constructor() {
