+++ test/fixtures/my-bin/lib/test_command.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../..').Command;
+const Command = require('../../../..');
 
 class TestCommand extends Command {
   constructor() {
