+++ test/fixtures/my-git/lib/command/remote.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../../..').Command;
+const Command = require('../../../../..');
 
 class RemoteCommand extends Command {
   constructor() {
