+++ test/fixtures/my-helper/index.js
