+++ test/fixtures/my-helper/lib/command/echo.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../../..').Command;
+const Command = require('../../../../..');
 
 class EchoCommand extends Command {
   constructor() {
