+++ test/fixtures/my-helper/lib/command/fork.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../../..').Command;
+const Command = require('../../../../..');
 const path = require('path');
 
 class ForkCommand extends Command {
